Remove FinalizedReport; render directly from Report
The finalize() seam deferred no real work: OperationMetrics already folds
every span into O(1) running stats at record time, so finalize() only
sorted operations by name and copied cheap getters into an owned struct.
Both stdout and JSON read the same deterministic getters on the same
immutable metrics, so no shared "finalized" snapshot is needed for
consistency.

Move stdout table rendering onto `impl Display for Report` and JSON
writing onto `impl Report`, sharing a private `sorted_operations()`
helper. Delete finalized.rs, drop the `.finalize()` hop from
`Session::drop`, and relocate the finalized.rs tests into report.rs.

Also drop the plain arithmetic mean from the machine-readable JSON output,
completing the "slope is the primary metric" consistency between the API
and the JSON: the underlying `OperationMetrics::mean*` getters are now
`#[cfg(test)]`.
