+++ packages/all_the_time/src/report.rs
@@ -5,7 +5,7 @@ use std::fmt;
 use std::time::Duration;
 
 use crate::statistics::{format_slope_nanos, nanos_to_duration};
-use crate::{FinalizedOperation, FinalizedReport, OperationMetrics, OperationStatistics};
+use crate::{OperationMetrics, OperationStatistics};
 
 /// Thread-safe processor time tracking report.
 ///
@@ -170,45 +170,36 @@ impl Report {
         }
     }
 
-    /// Fully computes this report into a [`FinalizedReport`].
+    /// Returns the operations sorted by name.
     ///
-    /// Both the stdout summary and the machine-readable JSON output are rendered
-    /// from the returned value.
-    #[must_use]
-    pub fn finalize(&self) -> FinalizedReport {
-        let operations = self
+    /// The report holds operations in an unordered map, so every output sorts
+    /// them by name to present a stable, reproducible order.
+    pub(crate) fn sorted_operations(&self) -> Vec<(&str, &ReportOperation)> {
+        let mut operations: Vec<(&str, &ReportOperation)> = self
             .operations
             .iter()
-            .map(|(name, operation)| {
-                FinalizedOperation::new(
-                    name.clone(),
-                    operation.total_iterations(),
-                    operation.total_processor_time(),
-                    operation.metrics.mean(),
-                    operation.statistics(),
-                )
-            })
+            .map(|(name, op)| (name.as_str(), op))
             .collect();
-
-        FinalizedReport::new(operations)
+        operations.sort_unstable_by_key(|(name, _)| *name);
+        operations
     }
 
     /// Prints the processor time statistics to stdout.
     ///
-    /// Finalizes the report and prints the resulting summary. Prints nothing if
-    /// no operations were captured. This may indicate that the session was part
-    /// of a "list available benchmarks" probe run instead of some real activity,
-    /// in which case printing anything might violate the output protocol the tool
-    /// is speaking.
+    /// Prints nothing if no operations were captured. This may indicate that the
+    /// session was part of a "list available benchmarks" probe run instead of
+    /// some real activity, in which case printing anything might violate the
+    /// output protocol the tool is speaking.
     // Excluded from coverage as an un-assertable stdout side effect, matching the
-    // sibling `Display` impls. The `finalize()` computation it performs is covered
-    // independently — `Session::drop` finalizes on the measured path and
-    // `FinalizedReport` is unit-tested — so nothing computational is hidden here;
-    // only the stdout emission, which cannot be meaningfully asserted, is skipped.
+    // sibling `Display` impl. The figures it prints are covered independently via
+    // `Display` and the JSON output, so nothing computational is hidden here.
     #[cfg_attr(coverage_nightly, coverage(off))]
     #[cfg_attr(test, mutants::skip)] // Too difficult to test stdout output reliably - manually tested.
     pub fn print_to_stdout(&self) {
-        self.finalize().print_to_stdout();
+        if self.is_empty() {
+            return;
+        }
+        println!("{self}");
     }
 
     /// Whether there is any recorded activity in this report.
@@ -330,15 +321,77 @@ impl fmt::Display for ReportOperation {
     }
 }
 
-// No API contract to test - output format is not guaranteed, and the rendering
-// itself is exercised through `FinalizedReport`.
-#[cfg_attr(coverage_nightly, coverage(off))]
+// No API contract to test - output format is not guaranteed.
+#[cfg_attr(coverage_nightly, coverage(off))] // Too annoying to test every question mark operator.
 impl fmt::Display for Report {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        // There is one report and it is always fully calculated: rendering the
-        // human summary goes through the same finalized value the JSON output
-        // uses, rather than a cheaper slope-only path.
-        write!(f, "{}", self.finalize())
+        if self.is_empty() {
+            writeln!(f, "No processor time statistics captured.")?;
+            return Ok(());
+        }
+        writeln!(f, "Processor time statistics:")?;
+        writeln!(f)?;
+
+        // Render the per-iteration slope, not the raw mean: the mean folds warmup
+        // and one-off costs into the figure, while the slope recovers the marginal
+        // per-iteration cost. The confidence interval is kept out of this summary
+        // for readability; it remains in the JSON output and the `statistics()`
+        // API. A "NaN" cell marks an operation whose spans covered zero iterations
+        // (an unusable measurement); "n/a" marks one that recorded no spans at all.
+        let cells: Vec<(&str, String)> = self
+            .sorted_operations()
+            .into_iter()
+            .map(|(name, operation)| {
+                let value = match operation.statistics() {
+                    Some(statistics) => format_slope_nanos(statistics.slope_nanos),
+                    None => "n/a".to_owned(),
+                };
+                (name, value)
+            })
+            .collect();
+
+        let max_name_width = cells
+            .iter()
+            .map(|(name, _)| name.len())
+            .max()
+            .unwrap_or(0)
+            .max("Operation".len());
+        let max_value_width = cells
+            .iter()
+            .map(|(_, value)| value.len())
+            .max()
+            .unwrap_or(0)
+            .max("Per iteration".len());
+
+        // Print table header.
+        writeln!(
+            f,
+            "| {:<name_width$} | {:>value_width$} |",
+            "Operation",
+            "Per iteration",
+            name_width = max_name_width,
+            value_width = max_value_width
+        )?;
+        let separator_name_width = max_name_width
+            .checked_add(2)
+            .expect("operation name width fits in memory, adding 2 cannot overflow");
+        let separator_value_width = max_value_width
+            .checked_add(2)
+            .expect("value width fits in memory, adding 2 cannot overflow");
+        writeln!(
+            f,
+            "|{:-<name_width$}|{:-<value_width$}|",
+            "",
+            "",
+            name_width = separator_name_width,
+            value_width = separator_value_width
+        )?;
+
+        // Print table rows.
+        for (name, value) in cells {
+            writeln!(f, "| {name:<max_name_width$} | {value:>max_value_width$} |")?;
+        }
+        Ok(())
     }
 }
 
@@ -380,6 +433,55 @@ mod tests {
 
         let report = session.to_report();
         assert!(!report.is_empty());
+    }
+
+    #[test]
+    fn report_with_registered_but_unmeasured_operation_is_empty() {
+        let session = create_test_session();
+        let _operation = session.operation("unmeasured");
+
+        let report = session.to_report();
+        assert!(report.is_empty());
+    }
+
+    #[test]
+    fn report_with_only_zero_iteration_spans_is_empty() {
+        // A span that covered zero iterations records a span (so statistics can be
+        // fit) but no measurable work, so the report is still empty. Guards against
+        // `is_empty` regressing to key off whether statistics exist.
+        let session = create_test_session();
+        {
+            let operation = session.operation("failed");
+            let _span = operation.measure_thread().iterations(0);
+        }
+
+        let report = session.to_report();
+        assert!(report.is_empty());
+        // The operation still recorded a span, so it exposes statistics.
+        let operations = report.sorted_operations();
+        let (_name, operation) = operations.first().expect("the report has one operation");
+        assert!(operation.statistics().is_some());
+    }
+
+    #[test]
+    fn operations_are_sorted_by_name() {
+        let session = create_test_session();
+        {
+            let operation = session.operation("zebra");
+            let _span = operation.measure_thread().iterations(1);
+        }
+        {
+            let operation = session.operation("alpha");
+            let _span = operation.measure_thread().iterations(1);
+        }
+
+        let report = session.to_report();
+        let names: Vec<&str> = report
+            .sorted_operations()
+            .into_iter()
+            .map(|(name, _)| name)
+            .collect();
+        assert_eq!(names, ["alpha", "zebra"]);
     }
 
     #[test]
