+++ packages/alloc_tracker/src/target_output.rs
@@ -6,25 +6,22 @@ use std::path::{Path, PathBuf};
 
 use serde::Serialize;
 
-use crate::FinalizedReport;
+use crate::Report;
 
 /// Subdirectory of the Cargo target directory that receives the JSON files.
 const OUTPUT_SUBDIRECTORY: &str = "alloc_tracker";
 
 /// Machine-readable allocation statistics for a single operation.
 ///
-/// Carries the per-iteration means plus the per-iteration slope with its
-/// confidence interval for each metric, mirroring the shape `all_the_time`
-/// writes for processor time. Interval fields are omitted when the interval
-/// cannot be estimated.
+/// Carries the per-iteration slope with its confidence interval for each metric,
+/// mirroring the shape `all_the_time` writes for processor time. Interval fields
+/// are omitted when the interval cannot be estimated.
 #[derive(Serialize)]
 struct OperationOutput<'a> {
     operation: &'a str,
     total_iterations: u64,
     total_bytes_allocated: u64,
     total_allocations_count: u64,
-    mean_bytes_per_iteration: u64,
-    mean_allocations_per_iteration: u64,
     span_count: u64,
     slope_bytes_per_iteration: f64,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -38,7 +35,7 @@ struct OperationOutput<'a> {
     interval_high_allocations_per_iteration: Option<f64>,
 }
 
-impl FinalizedReport {
+impl Report {
     /// Writes machine-readable JSON statistics into the Cargo target directory.
     ///
     /// One file is written per operation, named after the operation, at
@@ -91,8 +88,7 @@ impl FinalizedReport {
         // file name would otherwise silently overwrite each other's results.
         let mut file_names: HashMap<String, &str> = HashMap::new();
         let mut outputs: Vec<(PathBuf, String)> = Vec::new();
-        for operation in self.operations() {
-            let name = operation.name();
+        for (name, operation) in self.sorted_operations() {
             let Some(statistics) = operation.statistics() else {
                 // Registered but never measured operations have no spans and thus
                 // no statistics, so they leave no output file behind.
@@ -113,8 +109,6 @@ impl FinalizedReport {
                 total_iterations: operation.total_iterations(),
                 total_bytes_allocated: operation.total_bytes_allocated(),
                 total_allocations_count: operation.total_allocations_count(),
-                mean_bytes_per_iteration: operation.mean_bytes(),
-                mean_allocations_per_iteration: operation.mean_allocations(),
                 span_count: statistics.span_count,
                 slope_bytes_per_iteration: statistics.bytes.slope,
                 interval_low_bytes_per_iteration: statistics.bytes.interval.map(|(low, _)| low),
@@ -191,10 +185,7 @@ mod tests {
         let session = session_with_recorded_work("allocate_vec");
         let directory = tempfile::tempdir().unwrap();
 
-        session
-            .to_report()
-            .finalize()
-            .write_to_directory(directory.path());
+        session.to_report().write_to_directory(directory.path());
 
         let file = directory.path().join("allocate_vec.json");
         let value = read_json(&file);
@@ -215,18 +206,6 @@ mod tests {
             value.get("total_allocations_count").and_then(Value::as_u64),
             Some(8)
         );
-        assert_eq!(
-            value
-                .get("mean_bytes_per_iteration")
-                .and_then(Value::as_u64),
-            Some(200)
-        );
-        assert_eq!(
-            value
-                .get("mean_allocations_per_iteration")
-                .and_then(Value::as_u64),
-            Some(2)
-        );
         // A single recorded span yields a span count of one and per-metric slopes
         // equal to the per-iteration means, but no interval (a single span carries
         // no dispersion information), so the interval fields are omitted.
@@ -239,6 +218,9 @@ mod tests {
         );
         assert!(value.get("interval_low_bytes_per_iteration").is_none());
         assert!(value.get("interval_high_bytes_per_iteration").is_none());
+        // The raw means, standard deviation, minimum and maximum are not emitted.
+        assert!(value.get("mean_bytes_per_iteration").is_none());
+        assert!(value.get("mean_allocations_per_iteration").is_none());
         assert!(value.get("std_dev_bytes_per_iteration").is_none());
         assert!(value.get("min_bytes_per_iteration").is_none());
         assert!(value.get("max_bytes_per_iteration").is_none());
@@ -263,10 +245,7 @@ mod tests {
         }
         let directory = tempfile::tempdir().unwrap();
 
-        session
-            .to_report()
-            .finalize()
-            .write_to_directory(directory.path());
+        session.to_report().write_to_directory(directory.path());
 
         let value = read_json(&directory.path().join("allocate_vec.json"));
         assert_eq!(value.get("span_count").and_then(Value::as_u64), Some(2));
@@ -296,10 +275,7 @@ mod tests {
         }
         let directory = tempfile::tempdir().unwrap();
 
-        session
-            .to_report()
-            .finalize()
-            .write_to_directory(directory.path());
+        session.to_report().write_to_directory(directory.path());
 
         let value = read_json(&directory.path().join("failed.json"));
         // A zero-iteration measurement has no per-iteration rate; both slopes are
@@ -330,10 +306,7 @@ mod tests {
         let session = session_with_recorded_work("group/case name");
         let directory = tempfile::tempdir().unwrap();
 
-        session
-            .to_report()
-            .finalize()
-            .write_to_directory(directory.path());
+        session.to_report().write_to_directory(directory.path());
 
         let file = directory.path().join("group_case_name.json");
         assert!(file.exists());
@@ -352,7 +325,7 @@ mod tests {
         let directory = tempfile::tempdir().unwrap();
         let target = directory.path().join("nested");
 
-        session.to_report().finalize().write_to_directory(&target);
+        session.to_report().write_to_directory(&target);
 
         // Nothing is written, so the directory is not even created.
         assert!(!target.exists());
@@ -372,10 +345,7 @@ mod tests {
         let _unmeasured = session.operation("unmeasured");
 
         let directory = tempfile::tempdir().unwrap();
-        session
-            .to_report()
-            .finalize()
-            .write_to_directory(directory.path());
+        session.to_report().write_to_directory(directory.path());
 
         assert!(directory.path().join("measured.json").exists());
         assert!(!directory.path().join("unmeasured.json").exists());
@@ -389,10 +359,7 @@ mod tests {
         fs::write(&file, "stale contents").unwrap();
 
         let session = session_with_recorded_work("allocate_vec");
-        session
-            .to_report()
-            .finalize()
-            .write_to_directory(directory.path());
+        session.to_report().write_to_directory(directory.path());
 
         // Parsing succeeds only if the stale, non-JSON contents were replaced.
         let value = read_json(&file);
@@ -416,7 +383,6 @@ mod tests {
 
         session
             .to_report()
-            .finalize()
             .write_to_directory(blocker.join("nested"));
     }
 
@@ -430,10 +396,7 @@ mod tests {
         // A directory occupying the output file's path makes the file write fail.
         fs::create_dir_all(directory.path().join("allocate_vec.json")).unwrap();
 
-        session
-            .to_report()
-            .finalize()
-            .write_to_directory(directory.path());
+        session.to_report().write_to_directory(directory.path());
     }
 
     #[test]
@@ -453,7 +416,6 @@ mod tests {
         // never created.
         session
             .to_report()
-            .finalize()
             .write_to_directory("collision_is_detected_before_writing");
     }
 }
